refactor(root): moved header component to root level
