+++ src/components/Header/index.tsx
@@ -1,9 +1,7 @@
 import type { FC, ReactNode } from 'react'
 
-import { useUserContext } from '~/features/auth'
+import { AccountInfo, useUserContext } from '~/features/auth'
 import type { User } from '~/features/events'
-
-import { AccountInfo } from '../AccountInfo'
 
 import { SignIn } from './parts/SignIn'
 import { StyledHeader, StyledLogo } from './styled'
