+++ src/components/Header/parts/Logo/index.tsx
