+++ src/components/Header/parts/Logo/parts/LogoIcon/index.tsx
