+++ src/components/Header/parts/SignIn/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import * as theme from '../../../../theme'
+import { theme } from '~/features/ui'
 
 export const Anchor = styled.a`
   ${theme.typography.paragraph.small}
