+++ src/features/ui/index.ts
@@ -1,8 +1,6 @@
 export * from './components/AccessibleHidden'
-export * from './components/AccountInfo'
 export * from './components/Button'
 export * from './components/Container'
-export * from './components/Header'
 export * from './components/Input'
 export * from './components/StyleReset'
 export * from './components/VerticalCenter'
